Remove thread id from VanillaChronicle
VanillaChronicle was using thread id in the name of data files
written by the Appender.  Max thread id can be configured
under Linux (/proc/sys/kernel/pid_max) to be an arbitrary size,
which may not fit within the size allocated in the index entries.
To eliminate this possibility, this change replaces use of
thread id with an integer allocated at creation of an Appender.

- Thread id was replaced by an incrementing integer that is
  assigned to each VanillaAppender on creation
- 24 bits was allowed for the Appender id
- All thread aware code was removed from the VanillaChronicle
  so that client code is now responsible for ensuring that
  an Appender is only ever used by a single thread.

NOTE:  This introduces an incompatible behaviour change to VanillaAppender.
Previously each call to VanillaChronicle.createAppender() by the same thread
would return the same Appender instance.  Now each call to createAppender()
creates a new instance.  It is the responsibility of the caller to ensure
that an Appender instance is only ever used by a single thread.