Tailer test to show issue with backward reading cycles written long ago
I created a test to reproduce an issue with being unable to read from
the end of the queue if all of the following conditions are met:
1) The time is at least one cycle in the future compared to the time the
last update was written to the chronicle instance
2) We read FORWARD at least once after reaching end of the queue before
trying to read BACKWARD. In this test, queue has only one element, so we
are trying to read from it two times.

If both of this conditions are met, all further attempts to read
backward from the end of the queue are failing.